Add files via upload
Here is my completed R code. Had a lot of trouble importing the CSV file such that it could be used for the table and barplots. Ended up having to use the point-and-click method of importing the dataset, rather than writing the code for it. Tried to make one barplot with the three questions as clusters, but ultimately couldn't figure it out. Hope my code is helpful, looking forward to learning from what you guys wrote.